worker: Use new buffer_manager in pb protocol
