+++ worker/buildbot_worker/pb.py
@@ -117,25 +117,52 @@ class ProtocolCommandPb(ProtocolCommandBase):
             del args['workdir']
             del args['workerdest']
 
+    def _message_consumer(self, message):
+        # after self.buffer.append log message is of type:
+        # (key, (text, newline_indexes, line_times))
+        # only key and text is sent to master in PB protocol
+        # if message is not log, simply sends the value (e.g.[("rc", 0)])
+        for key, value in message:
+            if key in ['stdout', 'stderr', 'header']:
+                update = [{key: value[0]}, 0]
+            elif key == "log":
+                logname, data = value
+                update = [{key: (logname, data[0])}, 0]
+            else:
+                update = [{key: value}, 0]
+            updates = [update]
+            d = self.command_ref.callRemote("update", updates)
+            d.addErrback(self._ack_failed, "ProtocolCommandBase.send_update")
+
     # Returns a Deferred
     def protocol_update(self, data):
         # data is a list of tuples
         # first element of the tuple is dictionary key, second element is value
-        dl = []
         for key, value in data:
-            update = [{key: value}, 0]
-            updates = [update]
-            d = self.command_ref.callRemote("update", updates)
-            dl.append(d)
-        return defer.DeferredList(dl, fireOnOneErrback=True, consumeErrors=True)
+            if key in ['stdout', 'stderr', 'header']:
+                whole_line = self.split_lines(key, value)
+                if whole_line is not None:
+                    self.buffer.append(key, whole_line)
+            elif key == 'log':
+                logname, data = value
+                whole_line = self.split_lines(logname, data)
+                if whole_line is not None:
+                    self.buffer.append('log', (logname, whole_line))
+            else:
+                self.buffer.append(key, value)
+        return defer.succeed(None)
 
     def protocol_notify_on_disconnect(self):
         self.command_ref.notifyOnDisconnect(self.on_lost_remote_step)
 
-    # Returns a Deferred
+    @defer.inlineCallbacks
     def protocol_complete(self, failure):
+        d_update = self.flush_command_output()
         self.command_ref.dontNotifyOnDisconnect(self.on_lost_remote_step)
-        return self.command_ref.callRemote("complete", failure)
+        d_complete = self.command_ref.callRemote("complete", failure)
+
+        yield d_update
+        yield d_complete
 
     # Returns a Deferred
     def protocol_update_upload_file_close(self, writer):
